api/requirements.txt
langchain
langchain_community
langchain_openai
openai
tiktoken
qdrant-client
peewee==3.15.4
pypdf
pandoc
fastapi[all]
pyjwt
bcrypt
passlib
pytest
psycopg2-binary
pytest-asyncio
jose
httpx
pydantic
python-multipart
uvicorn
unstructured[all]
python-dotenv